refactor:expand env in tree
