+++ includes/execution.h
@@ -72,5 +72,9 @@ int		is_numeric(char *str);
 char	**copy_environment(char **original_env);
 char	*expand_env_var(char *var, char **envp);
 void	expand_env_vars_in_tree(t_tree *tree, char **envp);
+int		contains_env_var(char *str);
+void	append_prefix(t_expand_vars *vars, char *str, int i);
+void	append_remainder(t_expand_vars *vars, char *str, int i);
+int		process_variable(t_expand_vars *vars, char *str, char **envp);
 
 #endif
